Fixing starlark (WORKSPACE, BUILD) snippet creation
We did not use the correct comment character for starlark: `//` instead of `#`. And lines with `//` get formatted by the bazel formatter, which adds a space between @ and prefix:

- `// @prefix` => `// @ prefix`
- `// @description` => `// @ description`

We only see the modified prefix and description lines. This is why saving such snippets does not work. It prints the `@prefix is required`-warning when saving the snippet.

There was a problem with finding out which comment-block gets used for a language. It only considered vscode-builtin extension to supply such things. This patch removes this restriction and so allows loading from the 'phgn.vscode-starlark' extension. This one sets the correct comment-string `#`.